Added divya.apk, key.keystore and signed_jar.apk
Generated payload for android.
When we install divya.apk in a mobile and set receiving host in our metasploit then we will get the information of the mobile and can have access on that.
If we install raw bite directly, the anti-virus in mobile will detect it. So, for non-detection signed_jar(a key and signed certificate for divya.apk) is generated.